for-each-ref: do not output empty lines
If the format string expands to an empty string for a given ref, do not
print the empty line.

This is helpful when wanting to print only certain kinds of refs that
you can't already filter for.

For example, to exclude symbolic refs, use format string:
  "%(if)%(symref)%(then)%(else)%(refname)%(end)"

Or to include only symbolic refs, use:
  "%(if)%(symref)%(then)%(refname)%(end)"

Signed-off-by: Eric Freese <ericdfreese@gmail.com>